Move filter creation from AbstractMongodbQuery to AbstractFetchableMongodbQuery.
Filter creation requires query execution (for joins/DBRef's) hence it's depending on the actual execution method. AbstractFetchableMongodbQuery is using imperative API so that's the best fit in there.
